requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiosignal==1.4.0
annotated-types==0.7.0
APScheduler==3.11.0
async-timeout==4.0.3
attrs==25.4.0
Babel==2.9.1
certifi==2025.10.5
charset-normalizer==3.4.4
frozenlist==1.8.0
idna==3.11
magic-filter==1.0.12
multidict==6.7.0
propcache==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
pytz==2025.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
yarl==1.22.0
